+++ core/dbt/contracts/graph/components.py
@@ -1,7 +1,9 @@
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from datetime import timedelta
 from dbt.artifacts.resources import Time as TimeResource
-from typing import Dict
+from dbt_common.contracts.util import Mergeable
+from dbt_common.dataclass_schema import dbtClassMixin
+from typing import Dict, Optional
 
 
 @dataclass
@@ -15,3 +17,23 @@ class Time(TimeResource):
 
     def __bool__(self):
         return self.count is not None and self.period is not None
+
+
+@dataclass
+class FreshnessThreshold(dbtClassMixin, Mergeable):
+    warn_after: Optional[Time] = field(default_factory=Time)
+    error_after: Optional[Time] = field(default_factory=Time)
+    filter: Optional[str] = None
+
+    def status(self, age: float) -> "dbt.artifacts.schemas.results.FreshnessStatus":  # type: ignore # noqa F821
+        from dbt.artifacts.schemas.results import FreshnessStatus
+
+        if self.error_after and self.error_after.exceeded(age):
+            return FreshnessStatus.Error
+        elif self.warn_after and self.warn_after.exceeded(age):
+            return FreshnessStatus.Warn
+        else:
+            return FreshnessStatus.Pass
+
+    def __bool__(self):
+        return bool(self.warn_after) or bool(self.error_after)
